# Improve cold boot time of the lambdas