--- a/api-gateway-routes/api_gateway_proxy_function.rb
+++ b/api-gateway-routes/api_gateway_proxy_function.rb
@@ -39,10 +39,8 @@ def on_connect(event, context)
 
   # -- Create Lambda for this session --
   lambda_client = Aws::Lambda::Client.new(region: region)
-  api_endpoint = get_api_endpoint(event, context)
   payload = {
     :queueUrl => sqs_queue.queue_url,
-    :apiEndpoint => api_endpoint,
     :connectionId => request_context["connectionId"],
     :levelId => authorizer["level_id"],
     :options => authorizer["options"],