This is an update to move some expensive one-time steps to the initialization phase of the lambda. This allows us to take advantage of our [provisioned concurrency](https://docs.aws.amazon.com/lambda/latest/dg/provisioned-concurrency.html) in order to decrease our cold boot times.

On the Neighborhood and Console lambdas, cold boot times decreased from ~18s to ~9s
On the Theater lambda, cold boot times decreased from ~5s to ~3s

We may not be able to optimize our cold boot times much further than this because it appears that AWS Lambda only allows for a [10s initialization phase](https://docs.aws.amazon.com/lambda/latest/dg/runtimes-context.html). This is not widely publicized and may change. I haven't verified this myself, so it's worth checking if we can use more initialization time. If we can, we definitely should as we have another ~8s of time we can improve on the smaller lambdas.

Note: The theater lambda has more than 3x the RAM of the other two lambdas, and therefore is expected to execute faster.

Note: This only improves the "cold boot" time for the lambda. This is the time it takes to set up a new lambda container the first time it is used. Both before and after this change, the non-cold boot run time for the Theater lambda running a "Hello World" program is ~0.5s and for the other two is ~0.75s. 

As part of this change, I also went ahead and removed the disk space logs, which were very verbose. I kept the "disk space report" log which we use to track disk space issues in Cloud Watch.

As a follow-up change, I'd like to add some sort of time report, similar to the disk space report, which records the amount of time we spend doing various pieces of work so we can track performance metrics separately from the time the user's project runs. [JAVA-499](https://codedotorg.atlassian.net/browse/JAVA-499)

As an additional follow-up, we should look into using the Java "Ahead of Time" compiler to [optimize specifically for AWS Lambda](https://www.baeldung.com/ahead-of-time-compilation). [JAVA-497](https://codedotorg.atlassian.net/browse/JAVA-497)

We may also be able further improve performance by assigning provisioned concurrency to the auxiliary lambdas (such as the authorizer lambdas) and move some steps to their init phase as well. The cold boot time for all of these combined appears to be around 5s (from watching my wrist watch) whereas the later runs are closer to 1s. We may see less improvement from adding provisioned concurrency here since these lambdas are used by all project types and therefore are used more often and cold boot less often. [JAVA-500](https://codedotorg.atlassian.net/browse/JAVA-500)

Finally, here's a potentially helpful [slack conversation](https://codedotorg.slack.com/archives/C03CK49G9/p1647459220180329) that contributed to this change.